Add workaround for issue #19186
See https://github.com/ethereum/go-ethereum/issues/19186.

The root cause of this issue is a race condition in rpc/handler.go which
causes eth_calls to return null (encoded in hex as "0x0"). The
workaround, which was proposed by a community member on that issue page,
is simply to change the order of two lines of code.